requirements.txt
python-telegram-bot
requests
python-barcode
Pillow